DomainCreators return vector of initial refinement/extents
Since we cannot serialize DomainCreators getting the initial
refinement level or extents is not useful for a single block. Instead
a vector of refinement levels or extents is returned by the
DomainCreator. The integer ID of the Block must match with the
location of that Block's entries in the vectors.