Remove CASSERT
CASSERT does not work with GCC 5.  Compilation fails if any function
containing it is used in a constexpr context, so such functions are
effectively not constexpr and their CASSERTs can be replaced with
plain ASSERTs.

I have tried to judge in each case that used CASSERT whether it is
more important to have the function constexpr or to have the
assertion and made the appropriate change.