Switch FluxCommunication to Inertial frame for now
When we have dual frames we will need to have figured out how to deal
with retrieving Tags whose template parameters we don't know since
that ability will solve a lot of problems like "What frame is this
thing going to be in?" One solution to this problem is to have a base
class and use a tmpl::filter call over the list of tags to get
what should be the only tag deriving off the base class. To be
concrete, the AnalyticSolution tag can derive off of same
AnalyticSolutionBase tag, which is an empty struct and is only used to
remove the template parameters from the type. Another possibility is
to use matching on template-template parameters, though this can get
annoying since we usually pass the size as a non-type template
parameter and we don't have a syntax like:

template <?...>
struct blah;

where the '?' can be matched by either a type or non-type template
parameter.