Allow retrieving base tags from ConstGlobalCache
In some cases we want to be able to retrieve tags from the cache
without knowing the tags' template parameters. For example, say we
have a tag:

template <typename SolutionType>
struct AnalyticSolutionTag : SolutionBase {
  using type = SolutionType;
};

then when retrieving the analytic solution we don't know what the
`SolutionType` is, we just want whatever it is from the cache. What we
effectively want to do is template-parameter erasure, similar to type
erasure. This could be thought of as "partial type erasure". This
commit makes it so that

get<SolutionBase>(cache)

returns the analytic solution from the cache.