Correct metaexpression in DataBox::create_from
The old expression appears to have given tmpl::list<> if RemoveTags
was empty and AddTags otherwise.  The new expression gives the
intersection of AddTags and RemoveTags, matching the comment.  Due to
the way the result is used I believe this change had no actual effect
on the operation of the DataBox.