Fix CAPTURE_PRECISE for STL types
Catch (correctly) defines an operator<< in the Catch namespace, but
that hides operators in other namespaces.  For our custom types the
operator<< is in the same namespace as the type and so can be found
using ADL, but for our custom STL operators the operator<< is in a
different namespace (the global one) from all its arguments (std), and
so cannot be found.

This commit causes the lookup to occur in the global namespace (or
actually a detail namespace with no operator<< functions in it) so our
custom operators can be found.