Auto register simple actions with Charm++
This commit adds a factory-style mechanism to automatically register Actions
called via the simple_action method. The registration is done by having
a static member variable of a class generate a pointer array of base-class
pointers that point to derived classes with methods that register the
methods. This means the 'simple_actions' typelist in Parallel
Components is no longer necessary.