Generate shell script to run non-failure tests
On supercomputers where the tests are always launched through MPI/GNI/the
networking layer, we can only execute one networking executable per job launch,
which is not the way ctest runs the tests. Since we generally don't develop on
supercomputers, this isn't too inconvenient but we still want to be able to run
the majority of the tests fairly easily.