Add ability to do math with std::reference_wrapper
The changes here make it transparent whether a type in a math expression is a
`std::reference_wrapper<T>` or the underlying type itself without needing to
explicitly write the math operators for `T`. That is, types can inherit the math
ability and they will still work with reference_wrapper.