Add RunSingleTest executable
RunSingleTest allows one to compile only one of the many test files and SpECTRE
libraries, and run the tests within the file. This reduces turn around time when
editing low-level headers such as some of the ones found in `Utilities` that are
compiled into the pre-compiled header. The reason is that when the PCH changes
all libraries need to be recompiled, which is very time consuming and not
necessary if only editing something that affects `DataVector`, for example.