Remove Basis::lgl namespace, replaced by Spectral
Instead of calls to Basis::lgl use the Spectral functions now that come
in two versions: One is templated on the Basis and Quadrature, and the
other takes a Mesh<1>. The latter is preferred, so e.g. to retrieve
collocation points in dimension d call
Spectral::collocation_points(mesh.slice_through(d)).