All CoordMaps now have checked inverse functions.
inverse now returns boost::optional<>, which can be invalid
if one tries to inverse-map a point that is sufficiently outside
the range of the map, especially a point for which the map is
singular or not one-to-one.

Instantiations of inverse for DataVectors have been removed, since we
don't see a use case (outside of tests) for inverse acting on a DataVector.

Similarly, ElementMap::inverse is no longer used/tested for DataVectors.

Also, instantiations of inverse for std::reference_wrapper<const double>
have been removed, since we do not need these either.