Make DataBox non-copyable
Copying a DataBox just created aliases between all the members, which
resulted in the surprising situation that mutating one DataBox would
also change the other.  As a side-effect, this allowed const DataBoxes
to be mutated by creating a non-const aliasing version.

The only uses of DataBox copying were in three tests, two of which
were bugs and the third of which was verifying that copying a DataBox
did not break something and so is now irrelevant.