Disallow mutable aliasing from db::create_from
Allowing a DataBox produced via create_from to be mutated could
unexpectedly change the source DataBox, as described for the copy
constructor in the preceding commit.

This change forces all aliasing copies to be const temporaries (whose
lifetimes can be extended in the usual way), so their contents cannot
be mutated to change the state of the initial DataBox.

There were no non-trivial uses of the removed functionality.