Remove duplicate entries from boost::variant type
Boost only supports a certain number of entries in a variant by
default.  (I believe 20.)  This is apparently a limit on the number of
entries in the type list, not the actual number of distinct types in
the variant.  Since most of the DataBox types in an algorithm are the
same, we can reduce the length of the list significantly by
deduplicating.