Use AdvanceTime action to initialize TimeId
Instead of initializing Tags::TimeId and Tags::Next<Tags::TimeId> in
InitializeElement, we now just initialize Tags::Next<Tags::TimeId> to
the desired start TimeId and use an AdvanceTime call at the start of
the algorithm loop to set both tags to the correct values.  The
immediate benefit of this is that it removes the logic from
InitializeElement needed to handle stepping to a slab boundary.  In
the longer term, this should simplify the algorithm needed for
self-starting the time stepper.