Allow actions to retrieve component mocks
When using the mocking framework it needs to be possible to retrieve a proxy to
a component that is being mocked through the non-mock component. That is, if I
am mocking the component A with MockA then if an action calls
Parallel::get_parallel_component<A>(cache) it should get back a proxy to
MockA. This commit enables such behavior.