Add support for calling threaded actions remotely
Previously threaded actions could only be called on the local object, but we
want to be able to call them remotely. One case where this is useful is when
wanting to write data to disk on N nodes where N is less than the total number
of nodes being used. In this case we invoke a threaded action on node M which
starts writing to disk. At the same time nodes can still contribute their data
at different ObservationIds to the nodegroup element M using threaded
actions. Another use case is being able to queue a threaded action with the
runtime system so that we can relinquish control back to the runtime system and
let proper threading occur.