Start GrMhd executable
Adds Initialization structs for setting up an Evolution
These were factored from InitializeElement with the following changes:
- ConservativeSystem splits allocation and initialization
- Evolution does not initialize history with analytic data
- Added documentation about each initializer

Note that I haven't touched InitializeElement as others are actively
changing that for BCs and limiting.  Those changes should be
moved into the factored code and we can discuss what to do with
InitializeElement.