Don't let YAML exceptions escape the option parser
If the name of a specified option was empty converting it to a string
would fail with a BadConversion exception in the initial parse pass,
which I thought could only generate ParserExceptions.  Now we just
catch everything.

We still print out the parsing error message, which I think is not
technically correct because the input is valid YAML, but it's clearly
not a valid input file so this should be OK.

Fixes #887