add functors for standard and blaze math functions
Put into Utilities/Functional.hpp are macros for unary, binary, and binary
assign operators such that functors can easily be generated from a function
argument. This is then used to define functors associated with all supported
operations on DataVectors. For convenience in writing type-generic code,
step_function, invsqrt, and invcbrt are defined for primitives.

Also added the functionality for literal functionals for ease in creating more
intricate functionals, as well as increased documentation for functionals.