Add FixedHashMap, a compile-time max size hash table
A hash table where the buckets are not allocated on the heap for better
performance when dealing with a small number of elements (~10-20 at most). The
hash table is also capable of taking advantage of perfect hashes to make lookup
even faster and avoid hash collisions. The interface is the same as
std::unordered_map with a few lesser-used functions not implemented as well as
some that do not make sense for a stack-allocated hash table.