ParallelInterpolator: Remove Frame template parameter.
Add a Metavariables type alias `domain_frame` which is the Frame
of the Domain.  This eliminates a `Frame` template parameter on
all Actions (most of which don't use `Frame` at all, but needed
to have the template parameter so that they can call an action that
called an action that needs the Frame).

The only Frame needed is the Frame of the Domain, which should be the
same for all the InterpolationTargets because we have only one Domain.
If we want a compute_target_points that expects inputs in a different Frame,
then that compute_target_points Action will be responsible for knowing about
the additional Frame and supplying points in `domain_frame`; knowledge
of the additional Frame is never needed elsewhere.