ParallelInterpolator: Simplify occurrences of VolumeDim param.
Add a Metavariables constexpr static member `domain_dim`
which gives the dimension of the Domain.  This eliminates a VolumeDim
template parameter for many Actions, many of which don't use VolumeDim
but need it as a template parameter so that they can call an Action
that calls an Action that needs the VolumeDim.

The Actions that provide `compute_target_points` might still have a
VolumeDim template parameter.