Each Observer now defines reduction_data_tags.
This makes it easier to add different Observers to the same
executable (even if the Observers observe something of the same type
but different values).

In Observer/Helpers.hpp there are new metafunctions that
assist with building a list of reduction_data_tags and reduction_data
from a list of ReductionDatum, and one for making the list of
reduction_data_tags in Metavariables given a list of Observers.

Also changed Test_ObserveSurfaceIntegrals.cpp to call more than
one observer (including more than one that observes things of
the same type but different values---something that was difficult to
do before this commit).