Replace ObserveSurfaceIntegrals with ObserveTimeSeriesOnSurface.
Now the things being observed can be the result of arbitrary
ComputeItems (that return a type that is observable by WriteReductionData).
To do integrals, use the new ComputeItems that do the integrals.
So now it is possible, for instance, to observe things
like the Christodoulou Mass on an apparent horizon.

Also simplified some of the logic, eliminated hardcoding of doubles,
and eliminated the need for complicated indexing.