Rework the way executables are added
We want to be able to globally add properties and dependencies to executables.
For example, we want to be able to use something that will embed the source tree
into the executable or we want to be able use precompiled headers. Overriding
CMake's add_executable function sometimes causes segfaults, and so we wrap this
function to make sure we are using supported features of CMake when overriding
our own add_spectre_executable function.