Make temporal_id = Tags::TimeId for Interpolation
Previously, in all the Interpolation infrastructure and tests,
Metavariables::temporal_id was either TimeId or Time, depending
on the test (this was inconsistent between different tests).

But EvolveValenciaDivClean defines Metavariables::temporal_id
to be Tags::TimeId, which is neither a Time nor a TimeId, but is instead
a Tag whose type is a TimeId.

This change makes Interpolation consistent with EvolveValenciaDivClean,
by defining Metavariables::temporal_id to be Tags::TimeId everywhere.
This change also allows Interpolation to be used in the
EvolveValenciaDivClean executable (which will be done in a future commit)
because now both can use the same Metavariables::temporal_id.

This changes only NumericalAlgorithms/Interpolation and its tests:

 - Metavariables::temporal_id is now always ::Tags::TimeId
 - All function arguments formerly of type Metavariables::temporal_id
   are now of type Metavariables::temporal_id::type.
 - All internal maps with keys formerly of type Metavariables::temporal_id
   are now of type Metavariables::temporal_id::type.
 - In all cases where `Time`s were explicitly used as function
   arguments, indices of maps, or to compare temporal_id values, these
   have been replaced by `TimeId`s.  This occurred only in the tests;
   the main Interpolation code never explicitly used Time or TimeId,
   only Metavariables::temporal_id.

The changes touch many files, but are trivial.