Observe linear solver residuals
The linear solver residual monitor now contributes the residual to the
reduction observer for writing out to disk. The data is written into
the standard reduction file, but into a separate "/linear_residuals.dat" H5 group.

Also adds a phase to the linear solver algorithm tests to clean up the
written data files.