Use MAKE_GENERATOR macro where applicable
In tests where random numbers are generated using a generator seeded
by std::random_device, this commit replaces inline creation of the
seed and generator with a call to a (recently-added) macro. The
macro simplifies the code and provides uniform capturing of the seed
in case of a test failure.