Add tests of Minmod TCI
This is the first of two commits that splits the former Minmod
limiter tests into TCI tests and limiter tests.

This commit adds new tests of the Minmod TCI that are essentially
duplicated from the former limiter tests. A subsequent commit will
remove the equivalent tests from the limiter test.