Group linear solver input file options
This adds an option tag for the linear solver options and uses it
for the conjugate gradient and GMRES residual monitors. It allows
the options to appear in a "LinearSolver" group in the input file.
The options are stored in the global cache, but the convergence criteria
must also be present in the DataBox to be used by compute items.