AddTemporalIdsToInterpolationTarget now ignores old TemporalIds.
Previously it was envisioned that AddTemporalIdsToInterpolationTarget
would be called only once for every TemporalId, by some global mechanism.

But now we want AddTemporalIdsToInterpolationTarget to be called by
a NodeGroup ParallelComponent (namely Interpolator), so it will be
called many times (as opposed to only once) for each TemporalId.

This change allows AddTemporalIdsToInterpolationTarget to act as
before on the first call but to ignore duplicate calls, even if
some of the duplicate calls are so late that they arrive after
interpolation has been completed on the relevant TemporalId.