Fix LambdaPiN limiter over-activations
For a particular range of slopes, the minmod LambdaPiN limiter was
previously erroneously activating. The limiter would activate to
linearize the solution without reducing the slope.

This behavior was limited to slopes that were too big to avoid
limiting via the LambdaPiN troubled-cell indicator, but not so large
that the slope needed to be reduced. In this range of slopes, the
correct behavior of LambdaPiN is be to leave the solution unmodified
(in contrast, other minmod limiters linearize the solution even if
the slopes are not too steep).