Fix gcc 9 warnings and errors
There is a new -Wdeprecated-copy warning about implicitly declared
copy constructors and assignment operators in some cases.  (I'm not
actually sure these aren't false positives as there seems to be
strange interaction with inheritance going on, but it's easy to make
the compiler happy.)

The compiler seems to no longer accept __attribute__((always_inline))
on lambdas.

There's a new compiler bug involving lambdas and parameter packs:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=90538

Finally, a change to Variables was necessary to get gcc 7 working with
the DataVector changes.