Split flux communication and flux lifting scheme
This separates the flux communication code into the actions that do the communication
and the flux lifting scheme that defines what needs to be communicated. In particular,
this allows to communicate fluxes for different sets of variables, as needed for the
elliptic solver. It also allows to implement second-order flux lifiting schemes
in the future, which is planned for elliptic systems. The evolution code can also
migrate to this scheme.