Generalize mortars initializer
Moves the mortar initialization code from the Evolution/DG
initializer into a stand-alone action and adds a test.
This initializer is used by both the evolution and elliptic code.
Since it assumes that the initialization happens _before_ the first
timestep of the evolution or the first elliptic iteration, some
changes to the linear solver were necessary to make the elliptic
code compatible with that viewpoint. Also adds comments about this
in various places.