Simplify logic in MinmodTci
Instead of comparing doubles to determine if the minmod function
returned its first argument, just look at the `.activated` bool
provided in the minmod output for precisely this purpose.

Note: as a result of this change, the TCI is marginally less likely
to trigger, because there is no longer an 1e-8 fuzziness on the
check that the output matches the input. This should be irrelevant
in real applications.