Make linear solver initializer an action
This is the last remaining elliptic initializer to be converted to
actions, so `InitializeElement` can now be deleted. The new linear
solver initializer action can also be used to re-initialize the
linear solver, which allows to perform multiple linear solves in
sequence. This is needed for solving nonlinear equations, where
we need to solve the linearized equations in each step of the
nonlinear solver.