Clean up OptionTags for InterpolationTargetTag::compute_target_points.
Previously each InterpolationTargetTag was also the OptionsTag for
InterpolationTargetTag::compute_target_points (this was so you can have
many InterpolationTargetTags specified in the input file, with options
specified under the InterpolationTargetTag names, e.g. AhA or AhB).
This was moderately confusing because the InterpolationTargetTag struct
served two purposes. It became more confusing after the recent change
fdcb568 that made the ConstGlobalCache hold SimpleTags instead of
OptionTags; after that change, to get the Interpolator to work properly
with input files InterpolationTargetTag would need to be a
SimpleTag and an OptionTag, so one struct would serve 3 purposes.

This commit simplifies the above situation so that InterpolationTargetTag
no longer needs to tag options.

Now each InterpolationTargetTag::compute_target_points (LineSegment,
KerrHorizon, etc) has its own Tag and its own OptionsTag, separate
from InterpolationTargetTag.  To allow multiple InterpolationTargetTags
to be specified in the input file, these Tags and OptionTags are templated
on InterpolationTargetTag.