Simplify type aliases for ConstGlobalCache tags
This commit makes it easier to collect tags for objects in the
ConstGlobalCache from metavariables, components, and actions.  If no
tags are added by a metavariable, component, or action, the empty type list
can be omitted.  Also the collection of tags from actions in the
phase-dependent action list is done in helper metafunctions rather than
explictly done in each component.  The following changes were made:
- Rename const_global_cache_tag_list to const_global_cache_tags so there is
a consistent type alias in metavariables, components, and actions.
- const_global_cache_tags with empty lists are removed.

In addition, some clean up was done in ParallelComponentHelpers.