Check apply calls with const references
This could not cause wrong behavior, but caused a suboptimal error
message when a function was passed in that took a mutable, rather than
const, reference.

The bugfix is replacing `const std::add_lvalue_reference_t<T>` with
`const T&`.

Also generally remove the use of type modification metafunctions
because they are very verbose and somewhat error-prone.