Automatically dereference unique_ptrs in DataBox
If the type of a tag is std::unique_ptr<T>, creating and mutating it
use that type, but db::get and other read-only access instead give
just a T.  This matches the behavior of the const global cache.

db::item_type is split into db::item_type for the type passed to
create and db::const_item_type for type returned from db::get.