Cut down MinmodTci to only the TCI part
This removes some logic that should have been only part of the
Minmod limiter. For the most part, this commit removes logic and
simplifies interfaces and tests.

But as part of removing the Minmod-only logic, the bool returned by
the `troubled_cell_indicator` function can change in some scenarios.
This affects a few tests in Test_MinmodTci, and also the TCI
portions of Test_Weno.