Refactor WENO reconstruction interface
The new interface works for one DataVector at a time, where the old
interface worked on an entire Tensor at a time. The new interface
makes it easier to apply the WENO limiters to certain tensor
components only.

Note: this is an intermediate step of a larger refactor, and
introduces inefficiency by copying data into a temporary data
structure - this copy will be removed in the next steps of the
refactor.