ID AnalyticData for scalar waves in Kerr spacetime
Caveat: Does not provide dt<U>, as calculating those needs
calculating the jacobian from logical to inertial coordinates,
for which we need to set up the element mesh. And the inputs
to "variables" overloads here do no provide sufficient info to
set up the mesh, as element extents in different dimensions
are not known.

Warning: will work to start an evolution with a self-starting
time-stepper.