Prevent FPEs in BulgedCube::inverse map
The root function for finding the scaling factor handled the r==0 case
separately since the map is singular there. We know if we are at r==0 then we
are at the origin and we can short circuit the root find and just return {0,0,0}
as the logical coordinates. In clang9 there is an optimizer bug that resulted in
FPEs in the root function (I think similar bugs have been encountered before),
which is why this change is being made.