Move DomainCreator calls to Domain/Tags.cpp
Since a lot of files include Domain/Tags.hpp, not including all the domain
creators into Tags.hpp reduces what needs to be recompiled when the domain
creators are updated. This reduces the number of targets that need to be rebuilt
when any domain creator has its header file changed by ~120.