Don't use hdf5 module on wheeler.
The python/anaconda2-4.1.1 module also loads hdf5,
so it is not necessary to load an extra hdf5 module.

On the contrary, if you do load an hdf5 module and it is a
different version than anaconda's, you will get runtime errors.
The latter problem was noticed when the anaconda2-4.1.1 module was
changed but the hdf5 module was not.