Template ComputeTimeDerivative on computation struct
We would like to reduce compute tags and so moving the computation of the
divergence of the flux into ConservativeDuDt will be needed. However, we will
also need to be able to control how the divergence of the flux is done depending
on the numerical scheme being used. Since currently the computation struct is
retrieved from the system, and the system should not be templated on the
numerical scheme, it is not possible to have the RHS computation be different
for different numerical schemes. By having ComputeTimeDerivative take the RHS
computation as a template it will be easy to support things like DG-subcell
where the subcell could be various different FD or FV schemes.