Update test of Weno limiter
Previously, the Weno limiter test was responsible for testing that
the SimpleWeno and Hweno reconstructions did the correct work.
But with the recent additions of the simple_weno_impl and hweno_impl
helper functions, and their associated tests, the reconstruction is
now tested elsewhere in a more detailed and modular way.

This commit updates the Weno test to
- remove the reconstruction tests that are now done elsewhere
- add more detailed testing that the TCI triggers reconstruction
  when expected
- add tests that Weno with WenoType::SimpleWeno matches the output of
  simple_weno_impl, and Weno with WenoType::Hweno matches the output
  of hweno_impl. This is more a consistency check than a robust test
  of the implementation.