Add sanity check to constrution of 0d Meshes
It was previously possible to call the constructor of a 0d Mesh with
extents larger than 1, even though the resulting Mesh would only
have a single grid point. This commit adds a check that a 0d Mesh
is constructed with a single grid point, so that the constructor call
actually resembles the final Mesh.