Refactor SendPointsToInterpolator.
Currently there is a send_points_to_interpolator function that
needs to be refactored for the case when we interpolate without
an Interpolator parallel component.

Here we factor useful pieces of code out of
send_points_to_interpolator.  In a few commits from now, the function
send_points_to_interpolator will be removed in favor of a new
Action (which will be called SendPointsToInterpolator).