Add create_cxx_link_flag_target for CMake v3.13 and newer
In v3.13 CMake introduced INTERFACE_LINK_OPTIONS which can be used to propagate
linker flags the same way that INTERFACE_COMPILE_OPTIONS propagates compiler
flags. Adding a function and using it on newer versions of CMake so that when we
can require CMake v3.13 and up we just need to delete the old code.