SendPointsToInterpolator.hpp -> InterpolationTargetDetail.hpp
All this does is move code (and change includes).

There is no need to have a separate file called
SendPointsToInterpolator.hpp, especially since the function
send_points_to_interpolator that used to live there is gone,
and since there is now another file of the same name in a
different subdirectory.