Fix Hweno caching bug
This was a rare corner case bug that should not have affected any
real runs. To summarize:

Elements with a single neighbor should label that neighbor as the
primary neighbor for the Hweno fit, and there should be no neighbors
to exclude from the fit. The bug was in the construction of the
(should-have-been) empty list of neighbors to exclude: it incorrectly
contained one direction, with the default value of lower_xi.
The bug manifested itself when trying to limit on an element with a
single neighbor in lower_xi: that neighbor would simultaneously be
the primary neighbor for the fit AND ALSO be excluded.

The vector is now correctly empty.