Decouple elliptic system from linear solver
- Move linear solver variables from elliptic system to Metavariables
- Initialize linear solver variables by linear solver and all
  quantities related to the DG operator in another action

Details:

Some linear solvers such as GMRES and CG work by applying the linear
operator to an internal "operand" quantity, or even a "preconditioned
operand" when they support preconditioning, whereas others such as the
Schwarz smoother apply the linear operator to the system fields directly.
The elliptic system does not need to know that, so this change moves
the logic to the Metavariables that chooses which linear solver to use
and that defines the algorithm.