Fix iteration IDs in linear solvers
Iteration IDs now always start at zero, representing the initial state
of the algorithm, and increment by one when an iteration is complete.
This change makes the control flow much easier to understand and ensures
observations are consistent. Previous to the recently merged PR #1725
we had to keep track of the "next" iteration ID and deal with interactions
between incrementing the iteration ID and the flux communication, but
that complication is now removed.