InterpolationTargetReceiveVars: Remove use of deleted Action.
This fixes issue #2262

In earlier versions of the code, compute_target_points was an Action.
In PR #2180, commit 643e5de6f2, all of those Actions were removed
and replaced by a single Action called SendPointsToInterpolator.

However, one compute_target_points Action remained and was not
caught by the unit tests.

In this PR, InterpolationTargetReceiveVars now uses
SendPointsToInterpolator and the test is changed appropriately.