Add IWYU to container again
Restores IWYU to the container, with a recent release of IWYU
for clang-10. Previously, the container included IWYU for
clang-5, but this was removed when we dropped support for
clang-5.

Initial explorations suggest that the newer IWYU still produces many
of the same false positives that we had before, so we do not resume
checking IWYU compliance in CI at this time.