Fix accuracy loss in continued fraction expansion
Previously the expansion was terminated when a term could not be
determined due to floating point error buildup, even if it could be
determined that the term was significant.  Now known-significant terms
are always returned, even if their values are approximate.

Additionally, the error estimate was initialized to an incorrect value
for numbers significantly different from 1.