Add ElementCenteredSubdomainData for Schwarz solver
An element-centered subdomain consists of a central element and overlap
regions with all neighboring elements. This class holds data on such a
subdomain. It supports vector space operations (addition and scalar
multiplication) and an inner product, which allow to use this data type with
linear solvers (see e.g. `LinearSolver::Serial::Gmres`).