Remove TagsList template parameter from Subitems
It was unused except in Test_BaseTags where it does not need to be used.
The only potential use would be for inserting a subtag that is a db::BaseTag,
but we do not allow insertion via base tags for non-subitems. So I don't think
it makes sense to allow insertion of base subtags.