Remove unnecessary compute tags from linear solvers
The linear solvers used compute tags way too extensively. This commit
replaces most of the compute tags in the linear solvers with simple
function calls. As a result, also some of the simple tags can be
removed. This reduces the complexity of the code quite significantly.

Note that some of the bookkeeping simple tags on the residual monitor
singleton components are replaced by additional reduction data, i.e.
iteration IDs are sent along with the data. These few additional bytes
propably won't affect performance, since the cost is presumably
dominated by the communication overhead, not the amount of data sent.