Simplify Interface tags
Tags::Interface is now just a simple tag with no inheritence.  The DirectionsTag
and Tag template parameters must be simple tags as if they were compute tags they
would bypass the dependency graph of a DataBox.

With the recent changes to compute tags so that they are all mutating, the implementation
of InterfaceCompute has been simplified.  The DirectionsTag template parameter of Slice
and InterfaceCompute are restricted to be simple tags (this is independent of how the
DirectionsTag is inserted into the DataBox, it can still be inserted as a ComputeTag).