Allow locally specifying pypp conversion to python classes
Adds support for specifying how should be manipulated (e.g. get a single grid
point value out) before sending to python. This allows passing complicated
classes, such as an equation of state, into the python code without having to
write bindings or some other general complicated conversion class because the
equation of state can just be converted to a std::array<double> that holds the
values for the equation of state parameters.