Add Schwarz solver algorithm
This Schwarz-type linear solver works by solving many sub-problems in
parallel and combining their solutions as a weighted sum to converge towards
the global solution. Each sub-problem is the restriction of the global
problem to an element-centered subdomain, which consists of a central element
and an overlap region with its neighbors. The decomposition into independent
sub-problems makes this linear solver very parallelizable, avoiding global
synchronization points altogether. It is commonly used as a preconditioner to
Krylov-type solvers such as `LinearSolver::gmres::Gmres` or
`LinearSolver::cg::ConjugateGradient`, or as the "smoother" for a Multigrid
solver (which may in turn precondition a Krylov-type solver).