FocallyLiftedMap: Add source_is_between_focus_and_target.
Previously this parameter was a static member function of
the InnerMap.  But it turns out that the same InnerMap can
be used both for cases when the source is between the focus and
the target, and when it is not.  So now this parameter is a bool
passed into the constructor.