Include StlStreamDeclarations.hpp in MakeString.hpp
Removing the include of DataBoxTag.hpp in Domain/Tags.hpp removes a transitive
include of StlStreamDeclarations.hpp causing executables to fail to compile with
the error: "call to function 'operator<<' that is neither visible in the template
definition nor found by ADL" in the friend operator<< that returns by value.

This is a temporary fix until we fix includes and factor StdHelpers.hpp to avoid
including most of the STL when it StdHelpers.hpp is included.