Correct the compute tag that calculates the largest char speed
The purpose of ComputeLargestCharacteristicSpeed is to estimate
the largest characteristic speed in the bulk, such that the same
can be used to estimate the Courant factor when adjusting timestep
size. The current version could only do that for a given surface
along which normals are known. This commit changes that to get
the bulk maximum characteristic speed instead, which would
correspond to the maximum over all points in the bulk with an
optimally directed normal assumed at each point.